+++ src/components/about-me/index.tsx
@@ -1,1 +1,2 @@
-export { default as Summary } from "./Summary"
+export { default as Summary } from './Summary';
+export { default as Skills } from './Skills';
