+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { IconDiv, NavLink } from './nav';
 export { ProjectCard, ProjectCardDiv, DetailDiv } from './project';
+export { Skills, Summary } from './about-me';
