--- src/pages/Nav/index.tsx
@@ -1,2 +1,0 @@
-export { default as NavPage } from './NavPage';
-export { default as AboutMePage } from "./AboutMePage";
