+++ src/pages/index.tsx
@@ -3,5 +3,5 @@ import 정리용 페이지
 import {HomePage, NavPage, PageNotFound} from './components'    
 */
 export { HomePage } from './home';
-export { NavPage, ProjectPage, DetailPage } from './nav';
+export { NavPage, ProjectPage, DetailPage, AboutMePage } from './nav';
 export { PageNotFound } from './PageNotFound';
