+++ src/pages/nav/index.tsx
@@ -1,3 +1,5 @@
+// import 정리용 페이지
 export { default as NavPage } from './NavPage';
-export { default as ProjectPage } from './ProjectPage';
 export { default as DetailPage } from './DetailPage';
+export { default as ProjectPage } from './ProjectPage';
+export { default as AboutMePage } from './AboutMePage';
